GEODE-3256: Refactoring DataCommands
