+++ geode-core/src/main/java/org/apache/geode/management/internal/cli/commands/QueryCommand.java
@@ -18,6 +18,8 @@ package org.apache.geode.management.internal.cli.commands;
 import java.io.File;
 import java.util.Collections;
 import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
 import org.apache.commons.lang.StringUtils;
@@ -27,6 +29,8 @@ import org.springframework.shell.core.annotation.CliCommand;
 import org.springframework.shell.core.annotation.CliOption;
 
 import org.apache.geode.cache.CacheFactory;
+import org.apache.geode.cache.execute.FunctionService;
+import org.apache.geode.cache.execute.ResultCollector;
 import org.apache.geode.cache.query.QueryInvalidException;
 import org.apache.geode.cache.query.internal.CompiledValue;
 import org.apache.geode.cache.query.internal.QCompiler;
@@ -71,7 +75,7 @@ public class QueryCommand implements GfshCommand {
       return dataResult;
     }
 
-    Object array[] = DataCommands.replaceGfshEnvVar(query, CommandExecutionContext.getShellEnv());
+    Object array[] = replaceGfshEnvVar(query, CommandExecutionContext.getShellEnv());
     query = (String) array[1];
     boolean limitAdded = false;
 
@@ -97,7 +101,7 @@ public class QueryCommand implements GfshCommand {
       regionsInQuery = Collections.unmodifiableSet(regions);
       if (regionsInQuery.size() > 0) {
         Set<DistributedMember> members =
-            DataCommands.getQueryRegionsAssociatedMembers(regionsInQuery, cache, false);
+            DataCommandsUtils.getQueryRegionsAssociatedMembers(regionsInQuery, cache, false);
         if (members != null && members.size() > 0) {
           DataCommandFunction function = new DataCommandFunction();
           DataCommandRequest request = new DataCommandRequest();
@@ -107,7 +111,7 @@ public class QueryCommand implements GfshCommand {
           if (subject != null) {
             request.setPrincipal(subject.getPrincipal());
           }
-          dataResult = DataCommands.callFunctionForRegion(request, function, members);
+          dataResult = callFunctionForRegion(request, function, members);
           dataResult.setInputQuery(query);
           if (limitAdded) {
             dataResult.setLimit(CommandExecutionContext.getShellFetchSize());
@@ -128,5 +132,76 @@ public class QueryCommand implements GfshCommand {
       return DataCommandResult.createSelectInfoResult(null, null, -1, null,
           CliStrings.format(CliStrings.QUERY__MSG__INVALID_QUERY, qe.getMessage()), false);
     }
+  }
+
+  public static DataCommandResult callFunctionForRegion(DataCommandRequest request,
+      DataCommandFunction putfn, Set<DistributedMember> members) {
+
+    if (members.size() == 1) {
+      DistributedMember member = members.iterator().next();
+      ResultCollector collector =
+          FunctionService.onMember(member).setArguments(request).execute(putfn);
+      List list = (List) collector.getResult();
+      Object object = list.get(0);
+      if (object instanceof Throwable) {
+        Throwable error = (Throwable) object;
+        DataCommandResult result = new DataCommandResult();
+        result.setErorr(error);
+        result.setErrorString(error.getMessage());
+        return result;
+      }
+      DataCommandResult result = (DataCommandResult) list.get(0);
+      result.aggregate(null);
+      return result;
+    } else {
+      ResultCollector collector =
+          FunctionService.onMembers(members).setArguments(request).execute(putfn);
+      List list = (List) collector.getResult();
+      DataCommandResult result = null;
+      for (Object object : list) {
+        if (object instanceof Throwable) {
+          Throwable error = (Throwable) object;
+          result = new DataCommandResult();
+          result.setErorr(error);
+          result.setErrorString(error.getMessage());
+          return result;
+        }
+
+        if (result == null) {
+          result = (DataCommandResult) object;
+          result.aggregate(null);
+        } else {
+          result.aggregate((DataCommandResult) object);
+        }
+      }
+      return result;
+    }
+  }
+
+  private static Object[] replaceGfshEnvVar(String query, Map<String, String> gfshEnvVarMap) {
+    boolean done = false;
+    int startIndex = 0;
+    int replacedVars = 0;
+    while (!done) {
+      int index1 = query.indexOf("${", startIndex);
+      if (index1 == -1) {
+        break;
+      }
+      int index2 = query.indexOf("}", index1);
+      if (index2 == -1) {
+        break;
+      }
+      String var = query.substring(index1 + 2, index2);
+      String value = gfshEnvVarMap.get(var);
+      if (value != null) {
+        query = query.replaceAll("\\$\\{" + var + "\\}", value);
+        replacedVars++;
+      }
+      startIndex = index2 + 1;
+      if (startIndex >= query.length()) {
+        done = true;
+      }
+    }
+    return new Object[] {replacedVars, query};
   }
 }
